Remove a warning

+++ Combinatorial_map/test/Combinatorial_map/Combinatorial_map_copy_test.cpp
@@ -69,7 +69,8 @@ struct Map_3_dart_max_items_3
 
 struct MonInfo
 {
-  MonInfo(int i=0) : mnb(i==0?rand():i), ptr(reinterpret_cast<char*>(this))
+  MonInfo(long long int i=0) : mnb(i==0?rand():i),
+                               ptr(reinterpret_cast<char*>(this))
   {}
   int mnb;
   std::string s;
@@ -157,7 +158,7 @@ struct CreateAttributes
 {
   static void run(Map& map)
   {
-    int nb=0;
+    long long int nb=0;
     for(typename Map::Dart_range::iterator it=map.darts().begin(),
         itend=map.darts().end(); it!=itend; ++it)
     {
@@ -181,7 +182,7 @@ struct InitDartInfo
 {
   static void run(Map& map)
   {
-    long int nb=0;
+    long long int nb=0;
     for(typename Map::Dart_range::iterator it=map.darts().begin(),
         itend=map.darts().end(); it!=itend; ++it)
     {
